Improve saving/loading lock placement
- Create protocol should not show loading screen as soon as save dialog
  is open
- Save copy should not trigger save animation as soon as save dialog
  is open
