+++ src/other/dialogs.js
@@ -1,5 +1,26 @@
 import { remote } from 'electron';
 
+const openDialogOptions = {
+  buttonLabel: 'Open',
+  nameFieldLabel: 'Open:',
+  defaultPath: 'Protocol.netcanvas',
+  filters: [{ name: 'Network Canvas', extensions: ['netcanvas'] }],
+  properties: ['openFile'],
+};
+
+/**
+ * Shows a open dialog and resolves to (cancelled, filepath), which mirrors later
+ * versions of electron.
+ */
+const openProtocolDialog = () =>
+  new Promise((resolve) => {
+    remote.dialog.showOpenDialog(openDialogOptions, (filename) => {
+      const cancelled = filename === undefined;
+      const filePath = filename && filename[0];
+      resolve({ cancelled, filePath });
+    });
+  });
+
 const saveDialogOptions = {
   buttonLabel: 'Save',
   nameFieldLabel: 'Save:',
@@ -27,4 +48,9 @@ const saveProtocolDialog = (defaultPath = 'Protocol.netcanvas', saveCopy = false
     });
   });
 
-export default saveProtocolDialog;
+
+export {
+  saveProtocolDialog,
+  openProtocolDialog,
+};
+
