--- src/other/protocols/utils/__mocks__/openProtocolDialog.js
@@ -1,5 +1,0 @@
-/* eslint-env jest */
-
-const openProtocolDialog = jest.fn(() => Promise.resolve('/dev/null/fake/user/explored/path'));
-
-export default openProtocolDialog;
