--- src/other/protocols/utils/openProtocolDialog.js
@@ -1,24 +1,0 @@
-import { remote } from 'electron';
-
-const openDialogOptions = {
-  buttonLabel: 'Open',
-  nameFieldLabel: 'Open:',
-  defaultPath: 'Protocol.netcanvas',
-  filters: [{ name: 'Network Canvas', extensions: ['netcanvas'] }],
-  properties: ['openFile'],
-};
-
-/**
- * Shows a open dialog and resolves to (cancelled, filepath), which mirrors later
- * versions of electron.
- */
-const openProtocolDialog = () =>
-  new Promise((resolve) => {
-    remote.dialog.showOpenDialog(openDialogOptions, (filename) => {
-      const cancelled = filename === undefined;
-      const filePath = filename && filename[0];
-      resolve({ cancelled, filePath });
-    });
-  });
-
-export default openProtocolDialog;
